requirements.txt
absl-py==1.4.0
astunparse==1.6.3
bleach==6.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
et-xmlfile==1.1.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grad-cam @ git+https://github.com/jacobgil/pytorch-grad-cam.git@58a565a48fb3c28ac097f6a8533d554ac8ca2881
grpcio==1.56.0
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
imbalanced-learn==0.13.0
jax==0.4.13
Jinja2==3.1.2
joblib==1.3.1
kaggle==1.6.6
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
lit==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.6.1
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sklearn-compat==0.1.3
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.8.12
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
ttach==0.0.3
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.14.1
